stop trying to rewrite command history (nimitz)
This fixes the error that occurs when vosk_recognize_german tries to rewrite the command history.
Currently there does not seem to be any way to do so, so the feature was removed for now.

Background:
Since talon is in english command mode at the time the command is recognized, and the phrase is then run through the german engine, there are two history entries (one "nimitz <english garbled text>" and one "<some german phrase>").
Ideally these two entries would be overwritten with one entry "nimitz <some german phrase>", just like what you would get for the say command.