Implement cleaner solution for repeat count
This commit introduces a `repeat_count_small` capture that makes the definition of repeatable commands simple and elegant, removing the need for workarounds.

Previously I used a  workaround to get `repeat(number_small - 1)` to work when the user omits the number. This workaround not only was somewhat clunky,  aegis is "not particularly thrilled about that [if: repeat] being enabled by if due to the way repeat works".